V2 Update

+++ wire/modules/LanguageSupport/LanguageTranslator.php
@@ -3,7 +3,7 @@
 /**
  * ProcessWire Language Translator 
  *
- * ProcessWire 3.x, Copyright 2022 by Ryan Cramer
+ * ProcessWire 3.x, Copyright 2023 by Ryan Cramer
  * https://processwire.com
  *
  *
@@ -538,20 +538,32 @@ class LanguageTranslator extends Wire {
 		} else {
 			$textdomain = $this->filenameToTextdomain($filename);
 		}
-		$this->textdomains[$textdomain] = $this->textdomainTemplate(ltrim($filename, '/'), $textdomain); 
 		$file = $this->getTextdomainTranslationFile($textdomain); 
-		$result = file_put_contents($file, $this->encodeJSON($this->textdomains[$textdomain]), LOCK_EX); 
-		if($result && $this->config->chmodFile) chmod($file, octdec($this->config->chmodFile));
+		if(empty($file)) {
+			$this->error("Unable to get textdomain translation file: $textdomain");
+			return false;
+		}
+		if(is_file($file)) {
+			$result = true;
+		} else {
+			$this->textdomains[$textdomain] = $this->textdomainTemplate(ltrim($filename, '/'), $textdomain);
+			$result = file_put_contents($file, $this->encodeJSON($this->textdomains[$textdomain]), LOCK_EX);
+			if($result && $this->config->chmodFile) chmod($file, octdec($this->config->chmodFile));
+		}
 
 		if($result) {
 			$fieldName = 'language_files';
 			if(strpos($textdomain, 'wire--') !== 0)	{
-				if($this->wire('fields')->get('language_files_site')) {
+				if($this->wire()->fields->get('language_files_site')) {
 					$fieldName = 'language_files_site';
 				} 
 			}
-			$this->currentLanguage->$fieldName->add($file); 
-			if($save) $this->currentLanguage->save();
+			/** @var Pagefiles $pagefiles */
+			$pagefiles = $this->currentLanguage->$fieldName;
+			if(!$pagefiles->has(basename($file))) {
+				$pagefiles->add($file);
+				if($save) $this->currentLanguage->save();
+			}
 		}
 
 		return $result ? $textdomain : false;
@@ -632,6 +644,7 @@ class LanguageTranslator extends Wire {
 		$v = '';
 
 		switch(strtolower($str)) {
+			case '0-plural': $v = $this->_('0-plural'); break; // Is zero (0) plural or singular? type=radios options=[0-plural:Plural,0-singular:Singular] // i.e. would language use "0 items" (plural) or "0 item" (singular)?
 			case 'edit': $v = $this->_('Edit'); break;
 			case 'delete': $v = $this->_('Delete'); break;
 			case 'save': $v = $this->_('Save'); break;
@@ -669,6 +682,7 @@ class LanguageTranslator extends Wire {
 			case 'system': $v = $this->_('System'); break;
 			case 'modified': $v = $this->_('Modified'); break;
 			case 'error': $v = $this->_('Error'); break;
+			case ',': $v = $this->_(','); break; // Comma
 		}
 		
 		$level--;
@@ -676,5 +690,3 @@ class LanguageTranslator extends Wire {
 	}
 
 }
-
-
